+++ src/components/PeerList.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-    import { myName } from "@/stores/writeMyName";
-    import { peers } from "@/stores/writePeers";
-    import type { PeerRecord } from "@/stores/writePeers";
-    import type { APDesignerClientMessages } from "@/schemas/messages";
+    import { myName } from "#/stores/writeMyName";
+    import { peers } from "#/stores/writePeers";
+    import type { PeerRecord } from "#/stores/writePeers";
+    import type { APDesignerClientMessages } from "#/schemas/messages";
 
     const saveDisplayName = () => {
         // Broadcast your display name