+++ src/components/Pieces.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { state, type RenderRepModified } from "@/stores/writeState";
+    import { state, type RenderRepModified } from "#/stores/writeState";
     import { sheets } from "@abstractplay/renderer/src/sheets";
     import type { Glyph } from "@abstractplay/renderer";
     import { onMount } from "svelte";