+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2025-04-03
+
+### Added
+
+-   Added an animation feature. Create boards in the Play or Annotate tabs and click the camera icon to capture them. From the Animate tab, configure your frames and then generate an animated GIF.
+-   Added a named save/load feature for animation queues and for individual board positions. These are saved locally.
+
 ## [1.6.1] - 2025-01-13
 
 ### Added